+++ index.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
     <head>
-        <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
-        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <meta charset="UTF-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="index.css">
         <title>Awesome books </title>
@@ -32,6 +32,7 @@
                         <tr>
                         <th>AUTHOR</th>
                         <th>TITLE</th>
+                        <th></th>
                     </tr>
                     </thead>
                     <tbody class="book-List">
@@ -79,17 +80,18 @@
                 </div>
                 </section>
             
-                <section class="footer-section">
+                <div class="footer-section">
                     <div class="footer">
-                        <p>Copyright ...<br>
+                        <p> Copyright ...</p>
+                        
                     </div>
-                </section>
+                </div>
 
     </div>
 
 
 
-         <script type='modules' src="index.js"></script>
+         <script type="module" src="./modules/index.js"></script>
         <!-- <script src=""></script> -->
     </body>
 </html>
